+++ test/unit/fuzz-tools.test.js
@@ -324,6 +324,7 @@ import * as ficatArletV344 from '../../lib/ficat-arlet-v344.js';
 import * as lichtmanKienbockV345 from '../../lib/lichtman-kienbock-v345.js';
 import * as catterallPerthesV346 from '../../lib/catterall-perthes-v346.js';
 import * as herringPillarV347 from '../../lib/herring-pillar-v347.js';
+import * as strasbergBdiV348 from '../../lib/strasberg-bdi-v348.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -593,6 +594,7 @@ const MODULES = {
   'lichtman-kienbock-v345.js': lichtmanKienbockV345,
   'catterall-perthes-v346.js': catterallPerthesV346,
   'herring-pillar-v347.js': herringPillarV347,
+  'strasberg-bdi-v348.js': strasbergBdiV348,
   'tb-testing.js': tbTesting,
 };